Always populate TemplateParams 'Object'  (#249)
## Change Overview

This commit modifies template params creation to always
populate the Object field with the unstructured representation
of the Kubernetes object that is the target of the actionset

This allows writing generic blueprints that can be re-used
across different types of Kubernetes resources and to reference
fields from within the object spec that we do not hoist into
TemplateParams e.g. `.Object.metadata.labels.<foo>`

The change described is contained to `param.go` and `param_test.go`.
The rest of the changes are to the `kube` pkg to allow testing with
a fake dynamic client. 

## Pull request type

Please check the type of change your PR introduces:
- [ ] Work in Progress
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Trival/Minor
- [x] Bugfix
- [ ] Feature
- [ ] Documentation

## Issues

- #247 

## Test Plan

- [ ] Manual
- [x] Unit test
- [ ] E2E

```
$ go test -check.vv
...
OK: 14 passed
PASS
ok  	github.com/kanisterio/kanister/pkg/param	103.201s
```