Implementation for field.Field (#279)
This is a linked-list-based implementation. This is one of many possible approaches.

The approach attempts to prevent/limit the mutability of the fields outside the `field` package. Also, in the package the contents of `linkedField struct` are not modified after the struct has been created.

Include example usage (tests) to field package.

See #273