We ran this script in K10 with some specific dirs (#775)
which is not ideal, there are chances that some
files that we dont know of have these versions
now we just ignore mod,bin dirs and binary files
and running the script on the entire K10 repo.
We already introduced the issue by not running it
in the demos dir.