Bypass bash -c and run kando directly in the pod  (#798)
* Enclose kando --password into single quote

If we use the password as it is in the case of kand
location command and the password has some special char
for ex &, the operation fails.
This change tries to resolve that by replacing ' in the
password with \' and then enclosing the resultant into '.

* Enclose the command within " instead of ' and add
escape character before " and `

* Fix CI

* Address review comment

* Make kando command []string, remove enclosePassword

If we remove `bash -c` from the kando command the
command was string and thats why we faced problem
executing that.
When we remove bash -c and make the cmd []string
we are successfully able to run the thing withougt
the help of enclose password.

Co-authored-by: Pavan Navarathna <pavan@kasten.io>