Use CreateDBClusterSnapshot in case of Aurora for aws RDS
The backup mechanism as of now was db engine independent
and we simply used the API `CreateDBSnapshot` to create
the snapshot of the RDS instance that was provided in the
blueprint.
But in case of Aurora DB that seems to be a problem and
other API is suggested to be used.

```
InvalidParameterValue: The specified
      instance is a member of a cluster and a snapshot cannot be created
      directly. Please use the CreateDBClusterSnapshot API instead.
```
This PR tries to achive that by adding another optional argument in
the kanister function def.