Handle aurora db clusters in createRDSSnapshot and RestoreRDSSnapshot separately (#899)
* Use CreateDBClusterSnapshot in case of Aurora for aws RDS

The backup mechanism as of now was db engine independent
and we simply used the API `CreateDBSnapshot` to create
the snapshot of the RDS instance that was provided in the
blueprint.
But in case of Aurora DB that seems to be a problem and
other API is suggested to be used.

```
InvalidParameterValue: The specified
      instance is a member of a cluster and a snapshot cannot be created
      directly. Please use the CreateDBClusterSnapshot API instead.
```
This PR tries to achive that by adding another optional argument in
the kanister function def.

* Reword docs, fix docs CI

* Fix CI

* Fix document typo

* Support restore aurora db cluster, refactor

* Pass EngineVersion while calling RestoreDBClusterFromSnapshot, still failes

* Use poll.wait to wait for Aurora DB Cluster to be deleted and created successfully

* Fix CI, error checks

* Make debEngine optional in createRDSSnapshot and refactor

2. Update docs
3. Add blueprint for aurora db

* Generate cluster instance name from cluster name, and const instanceClass

* Refactor, minor comments

* re-trigger CI

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>