Remove `type` field from stable blueprints
Since we are upgrading the CRDs to `v1` from `v1beta1`, and we
don't actually specify `type` field in our `types.go` we wont
have `types` field in the CR schema. And because of that if
we create a blueprint with `types` in it, the validation would
failed.

I see there would be a problem when these applications are being
called from k10, because it looks like k10 needs type for `delete`
action.

I still am looking into if we really need that or we can change that.