Fix integration test failure by changing types.go
Even though e2e tests were running on my local cluster
tests were failing on the kanister CI, and the reason
was, when we read blueprint to create blueprint type
if a field is not specified in the blueprint, we initialized
the type with null for that field and in that case CRD validation
failed.
This commit tries to fix that by settign `omitemtpy` for those
fields.

I am thinking if we should add `omitempty` for every field in
types.go