Fix resource unit test failure
In the test `ResourceSuite.TestBlueprintClient` we tried to
create blueprint with empty actions struct and then tried
to make sure that we are able to update the blueprint. And
checked actions after update that its not nil.
Since we are omitting empty field we are not going to get
actions if we pass empty actions, that why I have changed
the action to have some values.