Make the AWS session token duration a configurable setting (#1067)
* Add assume role duration field to ExtractAWSCredentials.
Update callers of the function to pass default duration.
Update GetCredentials, so that it accepts the configured assume role duration instead
of the previous hard coded default.

* Added log for debugging purposes.

* Set the Duration field of the WebIdentityProvider object to control
the duration of the session token.

* Log expiration time.
Enable debug logging in AWS SDK

* Removed an info log.
Updated an info log to debug log to print the expiration time..

* Remove aws sdk log level config.
Add back log with assumeRoleDuration. But made it a debug log instead of info log.

* Update pkg/secrets/aws_test.go

Co-authored-by: Pavan Navarathna <pavan@kasten.io>

* Update comments for ExtractAWSCredentials.

* Fix a comment.

* Added helper function for getting creds with duration.
Added comments for the helper function.

Co-authored-by: Pavan Navarathna <pavan@kasten.io>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>