Update restore cmd in postgres depconfig blueprint
We recently figure out [an issue](https://github.com/kanisterio/kanister/pull/1116) where the restore command in our
blueprint was not working for postgres applications.
We fixed that for other application in the mentioned PR this
PR fixes the issue for the postgres application that gets
deployed as deployment config in OCP clusters.