Functionality to call ingress operations after discovering APIVersion (#1138)
* Figure out ingress API before calling List/Get operations

This commit adds wrapper around the std API's get and list
methods so that we can first figure out which api version
ingresses are part of, and then can call the respective methods.

* Add another method `IngressPath` in `IngressMgr` interface

Apart from getting/listting and creating the ingress resources
we would also need to get the ingress path of provided ingress
resource.
Since that would depend on which APIVersion the ingress resource
is from, its better we have that method as part of this interface.

* Add comments, make consts

* Address review comments

* Address review comment, change order of var init

* Address review comments

* Make ingress name format const

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>