Fix `gzip` not found issue on kanister-kubectl image
In the recent release (`0.70.0`) of kanister we changed the
base image of kanister-tools image that resulted into a failure
in kanister-kubectl image.
It started failing with `gzip` command not found, and the reason
is when we run `microdnf update` on previous rh base image it
install gzip but not in the newere rh base image.
This commit install gzip manually.