Change default value of `updateCRDs` to `true`
When we didn't have this `updateCRDs` flag in values.yaml
controller was responsible to create/update the CRDs. To
make sure that we follow the same default path we are changing
the default value of `updateCRDs` to `true`, so that kanister
controller will be responsible to manager crds. If thats
not required/possible, this flag can be set to false.