Fix `KubeExec` args in cassandra blueprint (#1277)
In the cassandra-blueprint for `KubeExec` function
in 2 phases we were using `containers` arguments which
is not supported by the function.
This has been working I think, because the container
defaults to the main container in the pod.
The is the reason instead of correct the typo
am removing it so that the behaviour, how it has
been runnig till now would not be changed.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>