Add helm test to check kanister upgrades (#1265)
* Add helm test to check kanister upgrades

This PR adds helm test that installs kanister from
released version then upgrades that to chart to
make sure we have not introduced any breaking chnages
and people will be able to upgrade without any issues.
We also use latest controller image when we upgrade
with local chart.

* Fix golangci-lint

* Minor refactor in function arguments

* Fix minor typos and add --create-namespace flag to helm install

* Add small comment while deleting NS

* Use `--create-namespace` instead of creating namespace manually

Instead of creating the namespace manually before installing
 a release, we can use `--create-namespace` flag in `helm
install` command that creates the namespace if not present.

* Refact test struct, run helm-test in another job

* Fix CI, integration-test

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>