Support an eventual phase for the blueprint actions (#1297)
* Support `deferPhase` in blueprint action

Blueprints can now define `deferPhase` for every action
that would be run once all (some in case of failure) the
phases of a blueprint action are run.

* Update docs for deferPhase changes

* Handle the scenario where only deferPhase is failed

* Try to fix CI by increasing actionset wait timeout

* Refactor code and add test for utility functions

* Address review comment, modularise functions

* Fix a bug while only deferPhase was failing

* Fix bug where actionset status was being set to complete even after
one of the core phases failed

The actionset status eventually must be failed if either one of core
phases or deferphases failed but we had a problem because of which
the actionset status was being set to complete even if the one of the
core phases failed. This commit fixes that.
It was happening because we were unknowingly setting error to nil
in deferPhase because deferPhase ran successfully.

* Address review comment, reword comments and docs

* Update the wait actionset timeout to 20 seconds

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>