Change format.Writer to not format phase output log lines
When kube.ExecOutput() writes the remote stdout/stderr outputs to the
user-provided writers, it calls format.Writer() to format the outputs
to the Kanister-log format. This causes the "phase output" to be
prepended with log metadata such as log timestamp, caller function etc.

A previous attempt to expand the output.Parse() function regex to handle
these log metadata prefixes has caused some downstream failures.

This fix reverts to the previous implementation of output.Parse(). It
also updates the format.Writer() to not format any "phase output".

Signed-off-by: Ivan Sim <ivan.sim@kasten.io>