Update log.entry() to not use global logger
Multiple log entries sharing a global logger can cause changes to the
entries' loggers to escape to the global level. For example, if an entry
is changed to output its log lines to a custom io.Writer, the global
logger will also be changed to write logs the custom io.Writer.

Also, when concurrent callers attempt to change the output target of
their respective log entries, the global logger will end up with racey
outputs.

This change clones the global logger inside the log.entry() function and
binds the clone logger to the entry instance.

Signed-off-by: Ivan Sim <ivan.sim@kasten.io>