Add `--egress-selector-mode` flag while starting k3d
Kanister CI tests were failing very frequently with below error while
execing a command into a pod

```
error dialing backend: EOF
```

We are not able to reproduce this, and did look like a networking
issue to me, because this was failing even before initiating the
pod deletion. Which means, it's not like the pod is not in the
correct state and that is the reason we see this failure.

By looking further I got to this issue https://github.com/k3s-io/k3s/issues/5633
which seems to be very closely related to our failure. This commit
incorporates the suggested change in our code.