Add a field in actionset status to figure out currently running phase (#1798)
* Add a field in actionset status to figure out currently running phase

This commit adds another field in the actionset status that can
be used to figure out the currently running phase of the actionset
set.

* Fix compilation error by passing nil actionset

* Add unit test for `actionset.status.progress.onPhase`

Add test to make sure the phases that are in blueprint are actually
being set in the actionset status.progress correctly.

* Change onPhase to runningPhase and reflect complete/failed

This commit changes the `onPhase` field to `runningPhase`
and also makes sure that `status.Progress.RunningPhase`
is updated with either complete of failed values, based
on the actionset's status, to make sure things are not
confusing.

* Running phase should be "" after actionset is completed

* Fix typo, correct error

* Update docs to reflect `Status.Progress.RunningPhase` change

* Improve docs sentence

* Address review comments, sentence/grammar improv

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>