Auto-generate CRD Documentation #1342  (#1727)
* created documentation

* Auto-generation of API CRD documentation

Added the command in make docs now so when make docs command is ran, the API CRD documentation is also created automatically with "API.md" name, created a folder called template which contains all the necessary files (template files, place-holder go file, json file and executable binary file)
which are used/needed in the creating of the API CRD documentation

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>

* Auto-Generation of API CRD documentation

Added the command in make docs now so when make docs command is ran, the API CRD documentation is also created automatically with "API.md" name, created a folder called template which contains all the necessary files (template files, place-holder go file, json file and executable binary file)
which are used/needed in the creating of the API CRD documentation

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>

* Deleting the directory

* adding template folder for internal PR

* Update Makefile

Co-authored-by: kale-amruta <41624751+kale-amruta@users.noreply.github.com>

* Modifying dokcer build image

Added tool to create CRD documentation to the docker build file, so we dont have to explicitly add exec binary to our project

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>

* Removing executable binary and json from template folder

As per the suggestion from the PR comment, removed executable binary from this folder, also removed .json file which was not required

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>

* Changes in makefile

Added a new target for creating on crd documentaion called as crd_docs and also added new command for it which is run inside the container image build

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>

* Update Makefile

Removed unwanted space

Co-authored-by: Prasad Ghangal <prasad.ghangal@gmail.com>

* Update Makefile

Improved the spacing as per the suggestion.

Co-authored-by: Prasad Ghangal <prasad.ghangal@gmail.com>

* Making it more tidy

Removed the template folder, updated API.md file, Made alignment changes

Signed-off-by: Kartik-Garg <kartik.garg@infracloud.io>

* Modifying makefile alignement

Modified space in makefile

Signed-off-by: Kartik-Garg <kartik.garg@infracloud.io>

* Made changes in the Dockerfile build and makfile

Added new required files for the API CRD documentation tool and made changes in the docker build and the makefile command to run the target to generate the documentation.
Changed config file name.

Signed-off-by: Kartik-Garg <kartik.garg@infracloud.io>

* Changes in CONTRIBUTING.md

Added steps to create the documentation when changes are made in Kanister API Types as well.

Signed-off-by: Kartik-Garg <kartik.garg@infracloud.io>

---------

Signed-off-by: Kartik-Garg <kartikgarg938@gmail.com>
Signed-off-by: Kartik-Garg <kartik.garg@infracloud.io>
Co-authored-by: kale-amruta <41624751+kale-amruta@users.noreply.github.com>
Co-authored-by: Prasad Ghangal <prasad.ghangal@gmail.com>