Change how we figure out container to get logs from
While running the kanister function, to read the artifact that the phase has
produced we get the logs of the kanister function pod (let's say `KubeTask`).
If there is just one container in the KubeTask pod in that case things would
work but if there are multiple contianers in the pod (let's say service mesh
injected one), in that case the logic that we have to figure out the container
name was not efficient. If was relying on the first container of the pod.

This commit changes that logic to read the container name from `podOptions` or
fallback to the default container name. This should work in all the cases, if
container name is not specified while creating `podOptions` we will use default
container name everywhere otherwise we will use the container name that is spec
ified in the podOptions.