chore(ci) leverage actions/setup-go cache to speed up CI
This avoids the step for creating, uploading and downloading
the artifacts.

The GH cache can be shared across jobs, workflows and
multiple executions of workflows, be it in the same PR,
separate PRs or branch push events.

This should be faster since caches are more easily reused
across jobs and worker instances.

This approach also requires less storage space, since the
cache can be easily reused across multiple executions, so
it does not require creating artifacts per workflow execution.