Use `0.17.0` version of `golang.org/x/net` in restic and kopia
While building the tools image, we were also building and packaging
`restic` and `kopia` with it.
Some of the direct/indirect deps of these projects had a vuln. because
of package `x/net`.
This commit updates the `go.mod` of these packages while building them
to make sure that we are using updated version `0.17.0` of that dep.