Fix CVE because of `golang.org/x/net` in `kanister-tools` image (#2419)
* Add `golang.org/x/net` in replace to make sure we use `v0.17.0` version

Some of the kanister's deps were using older version of `x/net` package which
resulted into some of the vuln. being reported in the image.
This commit makes sure that we are using 0.17.0 version of `x/net`.

* Use `0.17.0` version of `golang.org/x/net` in restic and kopia

While building the tools image, we were also building and packaging
`restic` and `kopia` with it.
Some of the direct/indirect deps of these projects had a vuln. because
of package `x/net`.
This commit updates the `go.mod` of these packages while building them
to make sure that we are using updated version `0.17.0` of that dep.

* Run go mod tidy and then commit changes