Fix the way action set name was being configured (#2430)
* Fix the way action set name was configred

`kanctl` was able to configure the actionset name using the flag called
`action-set`.
This has two problems
1. The better flag to give actionset name would be `--name` instead of
`--action-set`. Becuase there is another flag called `action`, so it
gets confusing between `--action-set` and `--action`.
2. Even though user would expect that if they have provided `--name`
the actionset would be named exactly that. But the problem was we
were using `generateName` in code that resulted in a string being
added in the name.

* Fix golint

* Fix typo

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>