Correct logic that waits for snapshot to be succeeded
We checked if the snapshot was succeeded by making sure that these two
conditions were false

- err!=nil
- provisioningState!=succeeded

second condition can be false even if the provisioning state is pending,
let's say. and because of that we would return success even in the case
of err being not nil and provisioningStaet being pending. Which would
be a problem.
This fixes that.