Implement method `SnapshotCreateWaitForCompletion` for `AdStorage` type (#2494)
* Implement method `SnapshotCreateWaitForCompletion`

The method `SnapshotCreateWaitForCompletion` for `AdStorage` was just returning
`nil`. Which caused some problems because we were not able to wait for snapshot
to complete, successfully.
This commit fixes that by implementing the mentioned method.

* Correct logic that waits for snapshot to be succeeded

We checked if the snapshot was succeeded by making sure that these two
conditions were false

- err!=nil
- provisioningState!=succeeded

second condition can be false even if the provisioning state is pending,
let's say. and because of that we would return success even in the case
of err being not nil and provisioningStaet being pending. Which would
be a problem.
This fixes that.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>