Improve `NewPodControllerForExistingPod` to set `podReady` to true (#2623)
* Improve `NewPodControllerForExistingPod` to set `podReady` to true

`NewPodControllerForExistingPod` function was supposed to return the
podController for an existing pod. If this function is caleld, it would mean
that the pod is already runnign and we are creating `podController` for that
pod.
But this function didn't set `podReady` field of podController to true that
was beign checked in some other methods to make sure that the pod is running.
And because of that those methods would see the pod to be not running and
complain.
This commit fixes that by making sure that the pod is running and sets
`podReady` to true.

* Improve comment to function