Set `cluster_name` var in logs vars separately and not in `init` (#2777)
If we set the `cluster_name` field in log variables (extran fields that are
printed in logs), inside the `init` function of `log` package, we would
un necessarily be calling k8s api even from the places where we don't have to
do anything with the k8s api.
To fix this, this commit introduces a new function that can be used to setup
that field in the logs vars and that function will only be called from the
places where we need the logs to have to the `cluster_name` field, for example
`kanctl`, `controller` and `repository-controller`.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>