Make development on macs easier
After this change goreleaser will build the `kanctl` binary that can be used
on arm64 machines and it will also build the controller container image for
correct platform so that it can be deployed on k8s clusters.
By default, I think it was building the container image for arm64 platform.