Fix the `make codegen` issue or permission denied
It looks like in the newer version of `code-generator` the script that
we call to generate code `generate-groups.sh` eventually also calls other
script `generate-internal-groups.sh` but we just make the first file
executable and because of that we were getting `permission denied` error
when we tried to run `make codegen`.

This commit fixes that by making sure that the other script is also
executable.