Fix a bug in the code where we merge actionset and bp labels/annotations
In the code where we are merging the labels/annotations passed via actionset
with the labels/annotations configured inthe blueprint we checked if tp.Annotations is nil
and then merged both the annotations. While doing this we introduced a bug, where if the
actionset labels and annotations are nil, and even though blueprint annotations are proivided
the kanister function created pod would not have the correct labels/annotations.

Even though this might not be a problem if the actionset is created by the kanctl command
it would be a problem if the pod labels and annotations set to nil by creating the actionset
resource yaml manually.

This commit fixes that.