Implement new method `Resolve` of kopia's `fs.Symlink`
We have a mock type in kanister that implements kopia's `fs.Symlink`
interface. Since this interface was recently modified to have a new
method the mock type that we have in Kanister started to not implement
that.
This commit makes sure that we add a dummy implementation of the
type.